build sign-in feature and patient profile

+++ lib/services/models/user_model.dart
@@ -1,7 +1,7 @@
 import 'dart:ffi';
 
 class User {
-  Long? id;
+  int? id;
   String? role;
   String? email;
   String? password;
